Added GPU matmul support
added GPU infrastructure and matmul on the gpu, other operations not added yet

+++ Duck/Functions/Value/Double/Add.cs
@@ -8,12 +8,12 @@ namespace Duck.Functions.Value.Double
 {
     public class Add : IDoubleValueFunction
     {
-        public static double Apply(double a, double b)
+        public static float Apply(float a, float b)
         {
             return a + b;
         }
 
-        public static (double a, double b) ApplyDerivative(double a, double b)
+        public static (float a, float b) ApplyDerivative(float a, float b)
         {
             return (1, 1);
         }
